removed redundant if statement & new test dir
